requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==22.1.0
Automat==0.8.0
autopep8==1.6.0
bcrypt==3.2.2
blinker==1.4
bottle==0.12.19
braceexpand==0.1.7
cachetools==5.0.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.4
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
discord-ext-bot==1.0.1
discord-py-interactions==4.0.2
discord-py-slash-command==3.0.3
discord.py==1.7.3
distro==1.4.0
entrypoints==0.3
Flask==2.0.2
frozenlist==1.3.3
google-api-core==2.7.1
google-api-python-client==2.40.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.55.0
gunicorn==20.1.0
httplib2==0.20.4
hyperlink==19.0.0
idna==2.8
import-expression==1.1.4
importlib-metadata==5.0.0
incremental==16.10.1
itsdangerous==2.0.1
Janome==0.4.2
Jinja2==3.0.3
jishaku==2.5.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
line-profiler==4.0.0
linecache2==1.0.0
MarkupSafe==2.0.1
more-itertools==4.2.0
multidict==6.0.2
mysqlclient==2.1.0
netifaces==0.10.4
oauthlib==3.1.0
orjson==3.6.3
pbr==5.4.5
pexpect==4.6.0
protobuf==3.19.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.8.0
pycparser==2.21
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.7
pyrsistent==0.15.5
pyserial==3.4
python-dotenv==0.19.2
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.8
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
ssh-import-id==5.10
testresources==2.0.0
toml==0.10.2
traceback2==1.4.0
Twisted==18.9.0
typing-extensions==4.1.1